# Add carry, sign, zero, and parity flags to integer division instructions