--- a/src/cpu/core_dyn_x86/decoder.h
+++ b/src/cpu/core_dyn_x86/decoder.h
@@ -1741,6 +1741,10 @@ static void dyn_grp3_eb(void) {
 		if (decode.modrm.mod==3)
 			gen_dop_byte(DOP_MOV,DREG(TMPB),0,&DynRegs[decode.modrm.rm&3],decode.modrm.rm&4);
 		gen_releasereg(DREG(EAX));
+
+		// The division helpers update the CPU test flags
+		set_skipflags(false);
+
 		gen_call_function((decode.modrm.reg==6) ? (void *)&dyn_helper_divb : (void *)&dyn_helper_idivb,
 			"%Rd%Drd",DREG(TMPB),DREG(TMPB));
 		dyn_check_bool_exception(DREG(TMPB));
@@ -1781,6 +1785,10 @@ static void dyn_grp3_ev(void) {
 		if (decode.modrm.mod==3)
 			gen_dop_word(DOP_MOV,decode.big_op,DREG(TMPW),&DynRegs[decode.modrm.rm]);
 		gen_releasereg(DREG(EAX));gen_releasereg(DREG(EDX));
+
+		// The division helpers update the CPU test flags
+		set_skipflags(false);
+
 		void * func=(decode.modrm.reg==6) ?
 			(decode.big_op ? (void *)&dyn_helper_divd : (void *)&dyn_helper_divw) :
 			(decode.big_op ? (void *)&dyn_helper_idivd : (void *)&dyn_helper_idivw);