--- a/src/cpu/flags.cpp
+++ b/src/cpu/flags.cpp
@@ -1061,7 +1061,7 @@ void set_cpu_test_flags_for_division(const T quotient) noexcept
 	lflags.type = t_DIV;
 
 	// Create a new test flags object with the flags we'll be settings
-	CpuTestFlags div_test_flags(FLAG_CF | FLAG_OF | FLAG_SF);
+	CpuTestFlags div_test_flags(FLAG_CF | FLAG_OF | FLAG_SF | FLAG_ZF);
 
 	// After performing IDIV, Intel's micro-operation clears the carry and
 	// overflow flags. Curiously, the 8086 documentation declares that the
@@ -1075,6 +1075,7 @@ void set_cpu_test_flags_for_division(const T quotient) noexcept
 	div_test_flags.has_carry        = false;
 	div_test_flags.has_overflow     = false;
 	div_test_flags.is_sign_negative = is_negative(quotient);
+	div_test_flags.is_zero          = (quotient == 0);
 
 	cpu_regs.ApplyTestFlags(div_test_flags);
 }