# Description

The existing integer division instructions (unsigned `IDIV*()` and signed `DIV*()`)  already set the overflow flag, however they weren't setting the other relevant flags (carry, sign, zero, and parity) or resetting the other flags back to their undefined state.

Presumably the majority of DOS applications and games out there just grab the resulting quotient and don't care about the flag bits.

Silmarils games use an unsigned 16-bit division and its resulting flags to calculate an attenuation value passed to OPL channel 1 across roughly 100 video frames. Every frame, the calculation (typically) produces a larger attenuation value until it reaches `0x3f`, which is the OPL chip's max channel attenuation (or silence).

**Without the division flags**

The game only sets the initial attenuation (0% or max volume) and final attenuation value (100% or silent) leaving the 100-frame loop ineffective - so the tone blasts at max volume for the roughly 1.5 seconds.

```
7.211 | OPL: Attenuating volume of channel 1 by 0.00%

1.5 seconds elapse with channel 1 at full volume without the attenuation loop

8.623 | OPL: Attenuating volume of channel 1 by 100.00%
```

https://github.com/dosbox-staging/dosbox-staging/assets/165201780/ac758a6d-ec7d-4207-8552-cf7ca5c89ff6


---

**With the division flags**

The game calculates and applies increasing attenuation:

```
6.167 | OPL: Attenuating volume of channel 1 by 1.59%
6.181 | OPL: Attenuating volume of channel 1 by 1.59%
6.196 | OPL: Attenuating volume of channel 1 by 3.17%
6.209 | OPL: Attenuating volume of channel 1 by 4.76%
6.223 | OPL: Attenuating volume of channel 1 by 4.76%
6.238 | OPL: Attenuating volume of channel 1 by 6.35%
6.252 | OPL: Attenuating volume of channel 1 by 6.35%
6.267 | OPL: Attenuating volume of channel 1 by 7.94%
6.281 | OPL: Attenuating volume of channel 1 by 9.52%
6.295 | OPL: Attenuating volume of channel 1 by 9.52%
6.310 | OPL: Attenuating volume of channel 1 by 11.11%
6.323 | OPL: Attenuating volume of channel 1 by 12.70%
6.338 | OPL: Attenuating volume of channel 1 by 12.70%
6.352 | OPL: Attenuating volume of channel 1 by 14.29%
6.366 | OPL: Attenuating volume of channel 1 by 15.87%
6.380 | OPL: Attenuating volume of channel 1 by 15.87%
6.394 | OPL: Attenuating volume of channel 1 by 17.46%
6.409 | OPL: Attenuating volume of channel 1 by 17.46%
6.423 | OPL: Attenuating volume of channel 1 by 19.05%
6.437 | OPL: Attenuating volume of channel 1 by 20.63%
6.453 | OPL: Attenuating volume of channel 1 by 20.63%
6.466 | OPL: Attenuating volume of channel 1 by 22.22%
6.480 | OPL: Attenuating volume of channel 1 by 23.81%
6.495 | OPL: Attenuating volume of channel 1 by 23.81%
6.510 | OPL: Attenuating volume of channel 1 by 25.40%
6.523 | OPL: Attenuating volume of channel 1 by 26.98%
6.538 | OPL: Attenuating volume of channel 1 by 26.98%
6.551 | OPL: Attenuating volume of channel 1 by 28.57%
6.567 | OPL: Attenuating volume of channel 1 by 28.57%
6.580 | OPL: Attenuating volume of channel 1 by 30.16%
6.595 | OPL: Attenuating volume of channel 1 by 31.75%
6.608 | OPL: Attenuating volume of channel 1 by 31.75%
6.623 | OPL: Attenuating volume of channel 1 by 33.33%
6.638 | OPL: Attenuating volume of channel 1 by 34.92%
6.651 | OPL: Attenuating volume of channel 1 by 34.92%
6.666 | OPL: Attenuating volume of channel 1 by 36.51%
6.681 | OPL: Attenuating volume of channel 1 by 36.51%
6.694 | OPL: Attenuating volume of channel 1 by 38.10%
6.708 | OPL: Attenuating volume of channel 1 by 39.68%
6.724 | OPL: Attenuating volume of channel 1 by 39.68%
6.737 | OPL: Attenuating volume of channel 1 by 41.27%
6.752 | OPL: Attenuating volume of channel 1 by 42.86%
6.765 | OPL: Attenuating volume of channel 1 by 42.86%
6.780 | OPL: Attenuating volume of channel 1 by 44.44%
6.794 | OPL: Attenuating volume of channel 1 by 46.03%
6.808 | OPL: Attenuating volume of channel 1 by 46.03%
6.822 | OPL: Attenuating volume of channel 1 by 47.62%
6.837 | OPL: Attenuating volume of channel 1 by 47.62%
6.851 | OPL: Attenuating volume of channel 1 by 49.21%
6.866 | OPL: Attenuating volume of channel 1 by 50.79%
6.880 | OPL: Attenuating volume of channel 1 by 50.79%
6.895 | OPL: Attenuating volume of channel 1 by 52.38%
6.909 | OPL: Attenuating volume of channel 1 by 53.97%
6.922 | OPL: Attenuating volume of channel 1 by 53.97%
6.937 | OPL: Attenuating volume of channel 1 by 55.56%
6.951 | OPL: Attenuating volume of channel 1 by 57.14%
6.965 | OPL: Attenuating volume of channel 1 by 57.14%
6.980 | OPL: Attenuating volume of channel 1 by 58.73%
6.995 | OPL: Attenuating volume of channel 1 by 58.73%
7.009 | OPL: Attenuating volume of channel 1 by 60.32%
7.022 | OPL: Attenuating volume of channel 1 by 61.90%
7.037 | OPL: Attenuating volume of channel 1 by 61.90%
7.052 | OPL: Attenuating volume of channel 1 by 63.49%
7.065 | OPL: Attenuating volume of channel 1 by 65.08%
7.079 | OPL: Attenuating volume of channel 1 by 65.08%
7.094 | OPL: Attenuating volume of channel 1 by 66.67%
7.108 | OPL: Attenuating volume of channel 1 by 68.25%
7.122 | OPL: Attenuating volume of channel 1 by 68.25%
7.137 | OPL: Attenuating volume of channel 1 by 69.84%
7.151 | OPL: Attenuating volume of channel 1 by 69.84%
7.165 | OPL: Attenuating volume of channel 1 by 71.43%
7.179 | OPL: Attenuating volume of channel 1 by 73.02%
7.194 | OPL: Attenuating volume of channel 1 by 73.02%
7.208 | OPL: Attenuating volume of channel 1 by 74.60%
7.222 | OPL: Attenuating volume of channel 1 by 76.19%
7.236 | OPL: Attenuating volume of channel 1 by 76.19%
7.251 | OPL: Attenuating volume of channel 1 by 77.78%
7.265 | OPL: Attenuating volume of channel 1 by 79.37%
7.279 | OPL: Attenuating volume of channel 1 by 79.37%
7.293 | OPL: Attenuating volume of channel 1 by 80.95%
7.308 | OPL: Attenuating volume of channel 1 by 80.95%
7.322 | OPL: Attenuating volume of channel 1 by 82.54%
7.337 | OPL: Attenuating volume of channel 1 by 84.13%
7.350 | OPL: Attenuating volume of channel 1 by 84.13%
7.366 | OPL: Attenuating volume of channel 1 by 85.71%
7.379 | OPL: Attenuating volume of channel 1 by 87.30%
7.394 | OPL: Attenuating volume of channel 1 by 87.30%
7.407 | OPL: Attenuating volume of channel 1 by 88.89%
7.423 | OPL: Attenuating volume of channel 1 by 88.89%
7.437 | OPL: Attenuating volume of channel 1 by 90.48%
7.455 | OPL: Attenuating volume of channel 1 by 92.06%
7.465 | OPL: Attenuating volume of channel 1 by 92.06%
7.479 | OPL: Attenuating volume of channel 1 by 93.65%
7.493 | OPL: Attenuating volume of channel 1 by 95.24%
7.508 | OPL: Attenuating volume of channel 1 by 95.24%
7.522 | OPL: Attenuating volume of channel 1 by 96.83%
7.536 | OPL: Attenuating volume of channel 1 by 98.41%
7.551 | OPL: Attenuating volume of channel 1 by 98.41%
7.564 | OPL: Attenuating volume of channel 1 by 100.00%
7.579 | OPL: Attenuating volume of channel 1 by 100.00%
7.579 | OPL: Attenuating volume of channel 1 by 100.00%
```

https://github.com/dosbox-staging/dosbox-staging/assets/165201780/01d58c44-20b8-460e-9236-4bad029adbfb

To inspect the OPL channel attenuation, save this to `log_opl_channel_attenuation.patch` then `git apply log_opl_channel_attenuation.patch` in the root of DOSBox Staging source tree:

```diff
diff --git a/src/hardware/opl.cpp b/src/hardware/opl.cpp
index b640ff79b..328a96dea 100644
--- a/src/hardware/opl.cpp
+++ b/src/hardware/opl.cpp
@@ -315,6 +315,13 @@ void Opl::WriteReg(const io_port_t selected_reg, const uint8_t val)
 
 
        } else { // OPL
+               if (selected_reg >= 0x40 && selected_reg <= (0x40 + 18)) {
+                       const auto opl_channel = 1 + selected_reg - 0x40;
+                       const auto attenuation_percent = (val * 100.0) / 0x3f;
+                       LOG_MSG("OPL: Attenuating volume of channel %d by %.2f%%",
+                               opl_channel,
+                               attenuation_percent);
+               }
                OPL3_WriteRegBuffered(&opl.chip, selected_reg, val);
                if (selected_reg == 0x105) {
                        opl.newm = selected_reg & 0x01;
```

## Related issues

Finishing touches on https://github.com/dosbox-staging/dosbox-staging/issues/3504

# Manual testing

Tested plenty of real mode, protected mode, and various Windows 3.1 programs and did not notice any problems or differences.

I suppose this could impact performance a bit, but I didn't see any (despite doing more work on the host). 


# Checklist

_Please tick the items as you have addressed them. Don't remove items; leave the ones that are not applicable unchecked._

I have:

- [x] followed the project's [contributing guidelines](https://github.com/dosbox-staging/dosbox-staging/blob/master/CONTRIBUTING.md) and [code of conduct](https://github.com/dosbox-staging/dosbox-staging/blob/master/CODE_OF_CONDUCT.md).
- [x] performed a self-review of my code.
- [x] commented on the particularly hard-to-understand areas of my code.
- [x] split my work into well-defined, bisectable commits, and I [named my commits well](https://github.com/dosbox-staging/dosbox-staging/blob/main/CONTRIBUTING.md#commit-messages).
- [ ] applied the appropriate labels (bug, enhancement, refactoring, documentation, etc.)
- [x] [checked](https://github.com/dosbox-staging/dosbox-staging/blob/main/scripts/compile_commits.sh) that all my commits can be built.
- [x] confirmed that my code does not cause performance regressions (e.g., by running the Quake benchmark).
- [ ] added unit tests where applicable to prove the correctness of my code and to avoid future regressions.
- [ ] made corresponding changes to the documentation or the website according to the [documentation guidelines](https://github.com/dosbox-staging/dosbox-staging/blob/main/DOCUMENTATION.md).
- [ ] [locally verified](https://github.com/dosbox-staging/dosbox-staging/blob/main/DOCUMENTATION.md#previewing-documentation-changes-locally) my website or documentation changes.