@weirddan455 2024-06-21
>  The existing integer division instructions (unsigned IDIV*() and signed DIV*()) already set the overflow flag, however they weren't setting the other relevant flags (carry, sign, zero, and parity) or resetting the other flags back to their undefined state.

I think this is incorrect, at least regarding the overflow flag.  A divide overflow is supposed to raise an exception.

Reference:  https://www.felixcloutier.com/x86/div

>  Overflow is indicated with the #DE (divide error) exception rather than with the CF flag.

Regarding flags:

>  The CF, OF, SF, ZF, AF, and PF flags are undefined.

I ran into this while writing some assembly testing my DOS file changes.  Here's a simple test case.  It can be assembled by `nasm hang.asm -o hang.com`  This program causes Staging to hang (not a full emulator hang but the shell becomes unresponsive).

```
org 100h

section .text
mov ax, 0ffffh
mov dx, 0ffffh
mov cx, 10
div cx

xor ax, ax
int 21h
```

In MS-DOS 6.22 the shell recognizes the overflow, displays an error, and resumes.  I was originally thinking this might just be a shell feature we don't have emulated but now I'm thinking this might be an inaccuracy with the CPU emulator.

![overflow](https://github.com/dosbox-staging/dosbox-staging/assets/8128047/e52ad75a-ed02-406f-b317-acb4073bf6d6)