@interloper98 2024-06-21
@weirddan455 , yes - the current code will actually generate an exception on division overflow. 

If it makes it past the exception check, then there's no overflow so clears the OF flag (this is what the original code did too).