@weirddan455 2024-06-21
>  yes - the existing/current code already generates an exception on division overflow (and that hasn't changed here).

Ok, I hadn't looked at the code yet.  I thought the issue I described might be related, but maybe it's not.  It looks like you've done your homework based on the comments in this PR.

>  Curiously, the 8086 documentation declares that the status flags are undefined following IDIV, but the microcode explicitly clears the carry and overflow flags.

So this game is depending on undefined CPU behavior, presumably.  Interesting.  Nice job finding this.

I really haven't worked on Dosbox's CPU emulator before so I'm maybe not the best person to review this but I'll give it a more detailed look later.