@interloper98 2024-06-21
Thanks @weirddan455 !
Any additional sanity checks and validation would be a huge help.