@Grounded0 2024-06-21
Here's what we use for performance profiling put together by myself:

https://anttipeltola.eu/misc/files/DOSBENCH.ZIP