@interloper98 2024-06-21
Adjusted the commits to sequence changes one flag at a time.

The first commit starts at functional-equivalence (just moved the same code into a function), then adds more flag handling commit-by-commit so we can isolate any problems to the exact flag (if this causes problems down the road).

Thanks @Grounded0. I'll run these before-and-after.