@johnnovak 2024-06-21
That's some standout work there @interloper98; an excellent demonstration of troubleshooting skills and the issue description is very well written 🚀 Quality work 💯 🧐 

I don't expect this to cause regressions, emulator-side performance problems, or compatibility issues because:

- You based your improvements on the actual reversed microcode, not on Intel's (likely incorrect) published documentation.
- Division is the absolutely slowest integer operation, so performance-conscious code avoided it like the plague. I expect division to be used rather sparingly in demos and games. Connsequently, I don't think this could cause measurable performance impact on the CPU emulation.
- As you've pointed out, 99.99% of DOS programs just don't care about these flags other than the overflow flag after a division. That makes sense to me. So we could even randomise those other flags and 99.99% of DOS programs would continue to work 😎 

I will give this a good testing later!