@interloper98 2024-06-22
Three runs each:

| Benchmark | menu key  | result units | main  | PR  |
|---|---|---|---|---|
| 3D Bench for fast PC | 2 | FPS | 368.9, 366.1, 368.36  | 369.7, 367.8, 369.7  |
| Chris's 640x480 | 4 | FPS | 332.0, 334.8, 324.8| 338.6, 335.2, 331.8|
| PC Player 640x480 | 6 | FPS | 282.3, 281.7, 282.6 |282.9, 281.3, 282.5 |
| Doom fast PC | b| realticks | 274, 278, 275| 276, 277, 279 |
| Quake 640x480 | e| FPS | 144.9, 144.7, 146.0| 145.9, 146.9, 145.8 |
| PerfDOS | p | FPS | 760 to 763 | 760 to 763 |
| Linkpack optimised | q | Mflops| 94.55, 94.14, 94.56 |95.10, 94.86, 94.19 |

Pentium MMX (w) lots of metrics, so took screenshots. Only one run each (results looked identical).

main:

![2024-06-21 17_34_00-](https://github.com/dosbox-staging/dosbox-staging/assets/165201780/870b10dc-07c9-4e74-8825-d09218e3e8bf)

PR:

![2024-06-21 17_46_23-DOSBox Status Window (Admin)](https://github.com/dosbox-staging/dosbox-staging/assets/165201780/4a2713ea-5928-4149-acb8-1224673c392c)