@johnnovak 2024-06-22
> Thanks for the comments @johnnovak;
> 
> > Division is the absolutely slowest integer operation, so performance-conscious code avoided it like the plague. I expect division to be used rather sparingly in demos and games.
> 
> The benchmarks results show that 100%! Cool; at least we're not weighing "correctness vs speed", in this case.

Good to see my theory backed up by hard data. Thanks for those measurements @interloper98 💯