@weirddan455 2024-06-22
The code looks good.  I think this should be a safe change, especially since you've researched how the real 8086 did it.

However, I'm pretty sure this will only affect the normal core.  I'm assuming the game you tested on is a real mode game and that's why this change worked.  The default core setting is `auto` which uses the normal core for real mode games and the dynamic core for protected mode games.

Those benchmarks you did are probably protected mode (I know for sure Doom and Quake are).  So they look identical because they are identical.  Not that I'm expecting setting a few flag bits to drastically affect performance though.

Have you tested this game with with `core = dynamic` to see if that is affected by the bug?  If so, you may also want to apply this fix to that core.  This doesn't have to be in this PR if you don't want to.  This fix is fine to go in as-is IMO.

For dyn_x86 see `src/cpu/core_dyn_x86/helpers.h`  -  This file contains only the division functions.  This is the default dynamic core used by x86 processors that is supposed to be highly optimized for the architecture.  I assume there is less translation that needs to happen since presumably some instructions can be near 1:1 map.

For the normal dynrec core see `src/cpu/core_dynrec/operators.h`  - Search for `dynrec_div_word`.  This is the core used by other architectures.  It can also be used on x86 but must be enabled at compile time by the Meson flag `-Ddynamic_core=dynrec` if you want to use it instead of dyn_x86.