@interloper98 2024-06-22
Thanks for catching these holes, @weirddan455!
Ideally I can reuse the same code for all; will get these filled so it's working right on all the cores.