@interloper98 2024-06-23
Thanks for the pointers into the dynrec and dynamic code paths, @weirddan455. Got those paths added now (wound up re-working the PR to use one function for all of them; wanted to avoid code duplication).

Tested each build type separately (`-Ddynamic_core=dyn-x86`, `-Ddynamic_core=dynrec`, `-Ddynamic_core=none`).

Re-checked performance a couple runs for each tests and the numbers are all in the same range as reported. (also checked normal an dnyrec cores and they're still on part w/ `main`).

Think its ready for review and any more testing the team can throw at it.